requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.1.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
Flask==2.2.5
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
numpy==2.0.1
openai==0.28.0
openpyxl==3.1.5
pandas==2.2.2
pandas-stubs==2.2.2.240603
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==3.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
types-pytz==2024.1.0.20240417
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==2.2.3
yarl==1.9.4
zipp==3.19.2
